Moved FEM tests to a dedicated project
