+++ ISAAR.MSolve.FEM.Tests/Resources/AbaqusFiles/instructions_for_Abaqus_jobs.txt
