+++ ISAAR.MSolve.FEM.Tests/Resources/AbaqusFiles/quad4_test0.inp
