+++ ISAAR.MSolve.FEM.Tests/Resources/AbaqusFiles/quad8_test0.inp
