+++ ISAAR.MSolve.FEM.Tests/Resources/AbaqusFiles/tri3_test0.inp
