+++ ISAAR.MSolve.FEM.Tests/Resources/AbaqusFiles/tri6_test0.inp
