+++ ISAAR.MSolve.sln
@@ -27,6 +27,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ISAAR.MSolve.Materials", "I
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ISAAR.MSolve.Geometry", "ISAAR.MSolve.Geometry\ISAAR.MSolve.Geometry.csproj", "{F1151C6B-C20E-4ED6-B2C3-7AA3D3EED508}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ISAAR.MSolve.FEM.Tests", "ISAAR.MSolve.FEM.Tests\ISAAR.MSolve.FEM.Tests.csproj", "{333B8164-88B7-409D-9A1E-1B8B11E3425D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -181,6 +183,18 @@ Global
 		{F1151C6B-C20E-4ED6-B2C3-7AA3D3EED508}.Release|x64.Build.0 = Release|Any CPU
 		{F1151C6B-C20E-4ED6-B2C3-7AA3D3EED508}.Release|x86.ActiveCfg = Release|Any CPU
 		{F1151C6B-C20E-4ED6-B2C3-7AA3D3EED508}.Release|x86.Build.0 = Release|Any CPU
+		{333B8164-88B7-409D-9A1E-1B8B11E3425D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{333B8164-88B7-409D-9A1E-1B8B11E3425D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{333B8164-88B7-409D-9A1E-1B8B11E3425D}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{333B8164-88B7-409D-9A1E-1B8B11E3425D}.Debug|x64.Build.0 = Debug|Any CPU
+		{333B8164-88B7-409D-9A1E-1B8B11E3425D}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{333B8164-88B7-409D-9A1E-1B8B11E3425D}.Debug|x86.Build.0 = Debug|Any CPU
+		{333B8164-88B7-409D-9A1E-1B8B11E3425D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{333B8164-88B7-409D-9A1E-1B8B11E3425D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{333B8164-88B7-409D-9A1E-1B8B11E3425D}.Release|x64.ActiveCfg = Release|Any CPU
+		{333B8164-88B7-409D-9A1E-1B8B11E3425D}.Release|x64.Build.0 = Release|Any CPU
+		{333B8164-88B7-409D-9A1E-1B8B11E3425D}.Release|x86.ActiveCfg = Release|Any CPU
+		{333B8164-88B7-409D-9A1E-1B8B11E3425D}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
